Create modules folder and reorganize files/imports
