+++ packages/core/src/js/helpers/index.js
@@ -3,6 +3,5 @@ export * from "./dispatchLifecycleHook";
 export * from "./buildCustomProps";
 export * from "./getPrefix";
 export * from "./getSetting";
-export * from "./pluginsArray";
 export * from "./setGlobalState";
 export * from "./setupPluginObject";
