+++ packages/core/src/js/modules/FocusTrap.js
