+++ packages/core/src/js/modules/eventEmitter.js
