+++ packages/core/src/js/modules/index.js
@@ -1,0 +1,4 @@
+export * from "./eventEmitter";
+export * from "./FocusTrap";
+export * from "./localStore";
+export * from "./PluginsArray";
