+++ packages/core/src/js/utilities/index.js
@@ -1,8 +1,5 @@
-export * from "./eventEmitter";
-export * from "./FocusTrap";
 export * from "./getConfig";
 export * from "./getElement";
-export * from "./localStore";
 export * from "./maybeRunMethod";
 export * from "./teleportElement";
 export * from "./toCamel";
