+++ packages/core/tests/modules/FocusTrap.test.js
@@ -1,6 +1,6 @@
 import "@testing-library/jest-dom/vitest";
 import userEvent from "@testing-library/user-event";
-import { FocusTrap } from "../../index";
+import { FocusTrap } from "../../src/js/modules";
 
 const user = userEvent.setup();
